pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cctvql"
version = "0.1.0"
description = "Conversational query layer for CCTV systems — ask your cameras anything in plain English"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "arunrajiah", email = "arunrajiah@gmail.com" }
]
keywords = ["cctv", "nlp", "surveillance", "frigate", "onvif", "llm", "conversational-ai", "nvr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Monitoring",
]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
mqtt = ["paho-mqtt>=2.0"]
onvif = ["onvif-zeep>=0.2.12"]
all = ["cctvql[mqtt,onvif]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest-cov>=5.0",
    "pre-commit>=3.7",
    "types-PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/arunrajiah/cctvql"
Repository = "https://github.com/arunrajiah/cctvql"
Issues = "https://github.com/arunrajiah/cctvql/issues"
Documentation = "https://github.com/arunrajiah/cctvql/wiki"

[project.scripts]
cctvql = "cctvql.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cctvql*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.coverage.run]
source = ["cctvql"]
omit = ["cctvql/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__",
    "if TYPE_CHECKING",
]
show_missing = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
